Allow header products to opt out of syncing
For the case where a module delegates the syncqt functionality to
another one, as it's done in the activeqt repository.
The QtHeaders instance of such a module would simply export a dependency
to the QtHeaders instance of the module that does the syncing, so
dependees won't have to know about the delegation.

Change-Id: Ide84b7ea2c910bc781a7b255f8afe8bdc79a6c35
Reviewed-by: Kai Koehne <kai.koehne@qt.io>